Refactor LevelCollecter to FileNameCollector as it's more accurate.
Add Tests, 100% Line coverage according to cobertura.
Also changed the class around a bit to be fully complete.
